add cublas version

+++ cuBLAS/Matrix/matmul/cpu_matrix.cpp
@@ -5,12 +5,12 @@
 using namespace std;
 
 template<>
-vector<int> MatrixMultiplication<cpu, int>::operator ()(const vector<int> &A, const vector<int> &B, int rA, int cA, int rB, int cB){
+vector<float> MatrixMultiplication<cpu, int>::operator ()(const vector<float> &A, const vector<float> &B, int rA, int cA, int rB, int cB){
 	clock_t t_start = clock(), t_end;
 	assert((int)A.size() == rA * cA);
 	assert((int)B.size() == rB * cB);
 	assert(cA == rB);
-	vector<int> C(rA * cB, 0);
+	vector<float> C(rA * cB, 0.);
 	for(int i=0;i<rA;++i) for(int j=0;j<cB;++j) for(int k=0;k<cA;++k) C[i*cB + j] += A[i*cA + k] * B[k*cB + j];
 	t_end = clock();
 	cout<<"CPU Matrix Multiplication Time Usage:\n";
@@ -19,9 +19,8 @@ vector<int> MatrixMultiplication<cpu, int>::operator ()(const vector<int> &A, co
 }
 
 
-/*
 int main(){
-	vector<int> A = {1,2,3,4,5,6}, B = {1,2,3,4,5,6};
+	vector<float> A = {1,2,3,4,5,6}, B = {1,2,3,4,5,6};
 	int rA = 2, cB = 2, cA = 3, rB = 3;
 	cout<<"A:"<<endl;
 	for(int i=0;i<rA;++i){
@@ -33,7 +32,7 @@ int main(){
 		for(int j=0;j<cB;++j) cout<<B[i*cB + j]<<' ';
 		cout<<endl;
 	}
-	vector<int> C = MatrixMultiplication<cpu, int>()(A, B, rA, cA, rB, cB);
+	vector<float> C = MatrixMultiplication<cpu, int>()(A, B, rA, cA, rB, cB);
 	cout<<"C:"<<endl;
 	for(int i=0;i<rA;++i){
 		for(int j=0;j<cB;++j) cout<<C[i*cB + j]<<' ';
@@ -41,4 +40,4 @@ int main(){
 	}
 	return 0;
 }
-*/
+
